Fix a flaky crash in Thread by guarding access to message_loop_ more carefully.
There was a window when message_loop_ could be non-NULL and invalid. The code
assumed that when message_loop_ is non-NULL, it's valid. Added a stopping_
flag which indicates a state in which we shouldn't access the message loop
because it is not safe.

Also, reused StopSoon logic in Stop and fixed a comment which was inaccurate.

TEST=Covered by base_unittests.
BUG=15331

Review URL: http://codereview.chromium.org/201108

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88f33312efe703fdb349afdcf4dc6d3c9147c033