Enable LeakTracker on release builds.
This is intended to be temporary, so we can get data from the dev channel on why URLRequest objects are leaking at shutdown.

WARNING: it may cause a slight slowdown on page cycler.

Also as part of this change, I switched the order that we check URLFetcher / URLRequest.
This is simply so that if both URLFetcher and URLRequest have leaked, we will report the URLFetcher leak rather than the URLRequest leak.

BUG=http://crbug.com/21199, http://crbug.com/18372

Review URL: http://codereview.chromium.org/217005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 03a69dcb6cbedb76522854b8bf6bde624d5c8301