Get rid of FilePath::AppendAndResolveRelative().
To resolve the problem of '..' parent references as well as symbolic links on POSIX platforms, we can simply use the file_util::AbsolutePath() function.

This has the drawback of having a different behavior on Windows and POSIX platforms, in the way that it can return a canonical path that doesn't exists when ran on Windows, but it will return an empty path (or false) when run on a POSIX platform.

So we need to add an extra PathExists() call to unify the behavior.

BUG=25681,25131 

Review URL: http://codereview.chromium.org/343003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a14b16bd4b6d7fc3c1e3e83f16d1d59bc05f3dab