Refactors RenderWidget to extract a PaintAggregator class.
After this change, I plan on changing the PaintAggregator algorithm.

Some things to note:

1- Previously, it was possible to send overlapping ViewHostMsg_PaintRect and
ViewHostMsg_ScrollRect messages.  This happened when scrolling a page since the
scrollbar would need to be repainted while the contents of the page are being
scrolled.  With this CL, this overlapping behavior is a bit more explicit.

2- There was a TODO about view_size clipping that I've eliminated.  I was able
to eliminate it because I realized that it is correct to clip the rects passed
by didInvalidateRect and didScrollRect to the size of the RenderWidget.
Apparently WebKit can sometimes invalidate regions outside the view.

R=brettw
BUG=25905
TEST=none

Review URL: http://codereview.chromium.org/403005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 552e600db9f51ac1efb0438bd749227226dffa64