Make SystemMonitor not a Singleton and move it out of base
SystemMonitor makes an assumption that through its lifetime a MessageLoop exists and stays the same. It is difficult and error-prone to satisfy that when it is a Singleton. It has caused problems in the past.

Additionally, extract HighResoltionTimerManager out of time_win.cc, eliminating yet another Singleton.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/431008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 097ae5a93a863b604fd6dbd939bed0c2be1afed2