Provide a quick and dirty way to reset about:objects data
To make it easier to use the about:object profiling facility,
I put in a quick/dirty way to reset all profile stats to 0 (as if
there were no births, deaths, etc.). This code is only activated
under debug builds (or if a developer inists in a private build).  
These stats don't impact semantics of the browser, so the hackish
approach to clearing the data counts can't instigate a crash, and 
it makes it much easier to look at changes in the stats.

While changing the code, I also added a lot of comments, and did
a few minor cleanups items.

I also officially added about:tasks as a replacement for about:objects,
as this is really how the service is used in Chrome. 

r=mbelshe
Review URL: http://codereview.chromium.org/100297

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 75b7920402c6f834376a18aa1927b223c6e3eadb