[Mac] address perf regressions on bots
This isn't the prettiest, but it should get the perf back where we want it.  If the file we're asked to exclude is in /var or /private, ignore it.  The problem on the bots is the TimeMachine prefs plist is growing unbounded as we keep adding 4 files each run for temporary profiles.  The load/modify/write gets slower with each addition.  This quick/dirty check to avoid that.  Since the original bug for the backup of this data is still open, I figure landing this is the right thing to do until that is fully addressed.

BUG=25959,30002
TEST=page cycles, browser_tests all get back to being fast.  normal profiles still exclude the files listed in the first bug.
Review URL: http://codereview.chromium.org/523024

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 254e043ce9545e1af39c76047b6264ae95c963cf