Auto restart when update available while running in the background on windows.
Landing on behalf of dbelenko@google.com.

This is a new version of the older patch. Main changes include: 1. Rebased to a never Git revision to make things easier to land. 2. Chrome will now preserve the command line switches except for those that are    blacklisted. 3. Fixed a race condition that would cause the browser to think it didn't exit    cleanly after it's been restarted. 4. Fixed minor nits and omissions (indentation, etc).

This patch adds a timer which fires every 6 hours and checks whether the
browser is in the "persistent" (background) mode, and whether there's an
update pending restart. If both conditions are true, the browser is
restarted with blacklisted command line keys and all loose values stripped. In order to
restart the browser in the background mode, the --long-lived-extensions
key is also added to the command line.

This change is Windows-only, and it won't become fully functional until
Drew (atwilson) checks in his work that enables Chrome to go into background.

Additionally, this addresses an issue where a restarted browser might load its profile data before the previous browser process exited - we now load the profile data *after* trying to contact the other browser. This exposed a race condition in SessionRestoreUITest.RestoreAfterClosingTabbedBrowserWithAppAndLaunching, so we disabled that test and logged http://crbug.com/40946.

BUG=40975,40946

Review URL: http://codereview.chromium.org/1618012

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd48c2b0bdcbb600bd428229c54cbefec46c5014