Acquire lock before iterating over list of field trials
The itteration is performed in the browser process
when the list of field-trial settings is passed to
the renderer.  I failed to acquire the lock prior
to the itteration, and it appears that some code
is deciding on a field trial VERY late in browser
startup and racing against this render-spawn
operation.  The code in FieldTrial is wrong
(but fixed here), but I'll still need to
track down the racing thread that is performing
such a late FieldTrial definition (the results
of which will not be passed to some renderers).

BUG=49409
R=huanr
Review URL: http://codereview.chromium.org/3134015

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43b8b5511efeff341d3855ef8c08056fcd30a621